Move swan forcing into a new module and several changes to docs
