+++ docs/source/core_concepts.rst
@@ -2,8 +2,9 @@
 Core Concepts
 =================================
 
-Rompy is a Python library for generating ocean model control files and required input data ready for ingestion into the model.
-The framework is separated into two broad concepts:
+Rompy is a Python library for generating ocean model control files and required input
+data ready for ingestion into the model. The framework is separated into two broad
+concepts:
 
 
 .. autosummary::
@@ -13,12 +14,14 @@ The framework is separated into two broad concepts:
     rompy.model.ModelRun
     rompy.core.BaseConfig
 
-There is information about each of these in the documentation of each object, but at a hight leve, is the high level framework 
-that renders the config object and controls the period of which the model is run, and the config object is responsible for producing 
+There is information about each of these in the documentation of each object, but at a
+high level, is the high level framework that renders the config object and controls the
+period of which the model is run, and the config object is responsible for producing
 model configuration. 
 
-If we consider a very simple case using the BaseConfig class. This is not inteded to do anything except provide a bass class on which to implement 
-a specific model, however, is is functional and can be used to demonstrate core concepts.
+If we consider a very simple case using the `BaseConfig` class. This is not inteded to
+do anything except provide a bass class on which to implement a specific model,
+however, is is functional and can be used to demonstrate core concepts.
 
 
 Core objects 
@@ -27,8 +30,8 @@ Core objects
 Grids
 ^^^^^
 
-Grids form a core component of any model. Rompy provides a base class for grids, and a regular grid class. 
-Support for other grid types will be added in the future.
+Grids form a core component of any model. Rompy provides a base class for grids, and a
+regular grid class. Support for other grid types will be added in the future.
 
 
 .. autosummary::
@@ -42,7 +45,8 @@ Support for other grid types will be added in the future.
 Data Objects
 ^^^^^^^^^^^^
 
-Data objects are used to represent data inputs into the model. Rompy provides a base class for data objects, and a data object for
+Data objects are used to represent data inputs into the model. Rompy provides the
+following base classes for data objects:
 
 .. autosummary::
     :nosignatures:
@@ -56,12 +60,8 @@ Data objects are used to represent data inputs into the model. Rompy provides a 
 Model Run 
 ---------------
 
-
 .. autosummary::
     :nosignatures:
     :toctree: _generated/
 
     rompy.model.ModelRun
-
-
-
