+++ docs/source/swan/components/group.rst
@@ -12,3 +12,4 @@ SWAN group commands
    rompy.swan.components.group.STARTUP
    rompy.swan.components.group.PHYSICS
    rompy.swan.components.group.OUTPUT
+   rompy.swan.components.group.LOCKUP
